backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pymongo==4.8.0
pydantic==2.9.1
pydantic[email]
python-dotenv==1.0.1
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
